refactor(#4): move address selection components into own folder
