+++ pages/five-days.tsx
@@ -7,10 +7,10 @@ import { useEffect, useState } from "react";
 import { Env } from "../src/api/env";
 import { maybeGetUser } from "../src/api/get-user";
 import type { ApiData, ApiData2 } from "../src/api/types";
-import AddressSearchInfo from "../src/components/common/AddressSearchInfo";
+import AddressSearchInfo from "../src/components/common/address/AddressSearchInfo";
+import SearchByAddress from "../src/components/common/address/SearchByAddress";
 import Heading2 from "../src/components/common/Heading2";
 import { colorWorkaroundGetServerSideProps } from "../src/components/common/layout/dark-mode-workaround";
-import SearchByAddress from "../src/components/common/SearchByAddress";
 import useGeolocationBasedWeather from "../src/components/common/use-geolocation-based-weather.hook";
 import WeatherBlock from "../src/components/common/weather/WeatherBlock";
 import { getLocationUrl } from "../src/components/get-location-url";
